fix: check forbidden errors

+++ src/routers/comments.router.ts
@@ -22,7 +22,13 @@ commentsRouter.get('/:id', async (req: RequestWithParams<{ id: string }>, res: R
 
 commentsRouter.delete('/:id', authTokenMiddleware, async (req: RequestWithParams<{ id: string }>, res: Response) => {
   const { id } = req.params;
+  const { userId } = req;
   if (!ObjectId.isValid(id)) return res.sendStatus(HTTP_STATUSES.NOT_FOUND_404);
+
+  //TODO: перенести это в сервис
+  const comment = await CommentQueryRepository.getCommentById(id);
+  if (!comment) return res.sendStatus(HTTP_STATUSES.NOT_FOUND_404);
+  if (comment.commentatorInfo.userId !== userId) return res.sendStatus(HTTP_STATUSES.FORBIDDEN_403);
 
   const isDeleted = await CommentService.removeCommentById(id);
   if (isDeleted) return res.sendStatus(HTTP_STATUSES.NO_CONTENT_204);
@@ -36,7 +42,13 @@ commentsRouter.put(
   async (req: RequestWithParamsAndBody<{ id: string }, IUpdateComment>, res: Response) => {
     const { id } = req.params;
     const { content } = req.body;
+    const { userId } = req;
     if (!ObjectId.isValid(id)) return res.sendStatus(HTTP_STATUSES.NOT_FOUND_404);
+
+    //TODO: перенести это в сервис
+    const comment = await CommentQueryRepository.getCommentById(id);
+    if (!comment) return res.sendStatus(HTTP_STATUSES.NOT_FOUND_404);
+    if (comment.commentatorInfo.userId !== userId) return res.sendStatus(HTTP_STATUSES.FORBIDDEN_403);
 
     const isUpdated = await CommentService.updateCommentById(id, { content });
     if (isUpdated) return res.sendStatus(HTTP_STATUSES.NO_CONTENT_204);
